downstream: KVM: arm64: vgic-v3: Upgrade AP1Rn to 64bit.
With the advent of FEAT_GIC_NMI, ICH_AP1R0_EL2 is now a 64bit
register, as the NMI priority is encoded in bit 63.

Upgrade the whole of the AP1Rn array to 64bit, leaving the
Group0 equivalent to 32bit.

+++ arch/arm64/kvm/hyp/vgic-v3-sr.c
@@ -130,7 +130,11 @@ static void __vgic_v3_write_ap0rn(u32 val, int n)
 	}
 }
 
-static void __vgic_v3_write_ap1rn(u32 val, int n)
+/*
+ * Contrary to ICH_AP0Rn_EL2, ICH_AP1R0_EL2 is 64bit, thanks to the
+ * NMI bit stuck at [63]. Isn't that fun?
+ */
+static void __vgic_v3_write_ap1rn(u64 val, int n)
 {
 	switch (n) {
 	case 0:
@@ -172,9 +176,10 @@ static u32 __vgic_v3_read_ap0rn(int n)
 	return val;
 }
 
-static u32 __vgic_v3_read_ap1rn(int n)
+/* Same remark about the 64bit-ness of AP1R0 */
+static u64 __vgic_v3_read_ap1rn(int n)
 {
-	u32 val;
+	u64 val;
 
 	switch (n) {
 	case 0:
